Added error handling and cancel function
